pyproject.toml
[tool.poetry]
name = "radiopribordatabase"
version = "0.1.0"
description = ""
authors = ["Igorik <egorsborsikov250575@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.40"
python-dotenv = "^1.1.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
